mysite/requirements.txt
asgiref==3.8.1
attrs==24.3.0
Django==4.2.14
django-filter==24.3
djangorestframework==3.14.0
docutils==0.21.2
drf-spectacular==0.28.0
flake8==7.1.1
flake8-docstrings==1.7.0
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
mccabe==0.7.0
pillow==11.0.0
pycodestyle==2.12.1
pydocstyle==6.3.0
pyflakes==3.2.0
pytz==2024.2
PyYAML==6.0.2
referencing==0.35.1
rpds-py==0.22.3
snowballstemmer==2.2.0
sqlparse==0.5.0
tomli==2.0.1
types-PyYAML==6.0.12.20240808
typing_extensions==4.12.2
uritemplate==4.1.1
